public/python/requirements.txt
flask
google-generativeai
pandas
nltk
scikit-learn